requirements.txt
# 投篮系统自动标定工具依赖

# 点云处理库
open3d>=0.18.0

# 数值计算
numpy>=1.24.0

# 配置文件解析
pyyaml>=6.0
